+++ src/Intcode/Instructions/IInstruction.cs
@@ -4,6 +4,6 @@ namespace Advent2019.Intcode.Instructions
 {
     public interface IInstruction
     { 
-        Task<int> RunAsync(Memory memory, int startingIndex, Command command, IInput input, IOutput output);
+        Task<long> RunAsync(Memory memory, long startingIndex, Command command, IInput input, IOutput output);
     }
 }
